Centralize data generation into one file
