+++ data/generation/utils.js
@@ -1,0 +1,20 @@
+const tab = str => `${str}\t`;
+
+const randomInt = (min, max) => min + Math.floor(Math.random() * (max - min));
+
+const ltrs = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j'];
+
+const generateName = (index, length) => {
+  let result = '';
+  for (let i = 0; i < length; i++) {
+    result += ltrs[index % 10];
+    index = Math.floor(index / 10);
+  }
+  return result;
+};
+
+module.exports = {
+  tab,
+  randomInt,
+  generateName,
+};
