+++ data/postgres-schema.sql
@@ -5,13 +5,10 @@ CREATE TABLE products (
   id SERIAL PRIMARY KEY,
   name VARCHAR(200) NOT NULL,
   brand VARCHAR(100) NOT NULL,
-  tier VARCHAR(50) NOT NULL,
-  -- options JSON NOT NULL,
-  -- price JSON NOT NULL,
-  -- about JSON NOT NULL,
+  product_tier VARCHAR(50) NOT NULL,
+  price INTEGER NOT NULL,
   is_prime BOOLEAN NOT NULL,
   stock_count INTEGER NOT NULL,
-  -- reviews JSON NOT NULL,
   questions INTEGER NOT NULL,
   seller VARCHAR(150) NOT NULL,
   thumbnail TEXT NOT NULL
